streamlit/requirements.txt
streamlit==1.39.0
pandas==2.3.3
psycopg2-binary==2.9.10
plotly==5.24.1